Move classes out of __init__ file, closes #2
