+++ tests/interval/test_coercion.py
