+++ tests/interval/test_initialization.py
