+++ tests/interval/test_operators.py
