+++ tests/interval/test_properties.py
