+++ tests/interval/test_representations.py
